Bluetooth: Runtime Bluetooth Interface Switch
This patch allows the bluetooth interface change at runtime.

The bluetooth interface defined by the Android property
"bluetooth.interface" will be used as bluetooth device.
ex: to use hci1
setprop bluetooth.interface hci1

If bluetooth is already when changing interface, a bluetooth
off/on is necessary.
If the property is not defined or malformed, interface hci0
is used by default.

Change-Id: I3a9c472f9313e4cc9048f12ac3bb30da5c88d7b1
Signed-off-by: lpoulain <loic.poulain@intel.com>