libbt-vendor: Add HCI device detection mechanism
In some cases, the HCI device could be not yet created.
This is the case when a serial BT controller is attached/detached
dynamically on BT power off/on (firmware loading can takes a while).
We need to detect dynamically the interface creation.

Use the Bluetooth management socket (hci control channel) to:
- Retrieve the interfaces list.
- If our interface does not exist, monitor interface creation events.

Use the HCIDEVDOWN dedicated ioctl to ensure the interface is down
before being able to bind with the hci user channel.

Change-Id: If2a0ef4859a39a04770c88abb8f6331f613c7940
Signed-off-by: Loic Poulain <loic.poulain@intel.com>